Fix reconcile handling when receiving events for Secrets and Configmaps
Two mapper methods are used. One for secrets and another for configMaps.
These mapper method converts the respective secret and configMap object
to KbsConfig and queues the reconciliation.

The reconcile loop now gets called with the KBSconfig object on any
change to secrets or configMaps in the operator namespace.

Additionally update the godeps/golang version as needed by the library
methods

Fixes: #20

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>