Set num of actix-web workers to 4 by default
When not set explicitly, the num of HTTP workers spawn by trustee is equal to the number of logical CPU cores available on the machine.
On systems with a high core count and hyper-threading enabled, this can result in an extremely large number of workers.
(in some cases, the "Too Many Open Files" system error can be hit)

Signed-off-by: Leonardo Milleri <lmilleri@redhat.com>